Moved colorinc/dec ROM_M_37: 6637 → 6516 - diff:  121 ROM_M_3A: 7718 → 7813 - diff: -95
Total size reduction: 26 bytes
