--- SAUCE/menustates/misc_xcd_bank_06.c
@@ -1,21 +1,0 @@
-// Will be merged into customize_screen.c
-
-void colorinc() {
-	if (idx8_inc(icon_colors, settingvalue) & 0x30) {
-		if ((uint8_t)(icon_colors[settingvalue] & 0x0F) >= 0x0D)
-			idx8_store(icon_colors, settingvalue, (icon_colors[settingvalue] + 0x10) & 0x30);
-	} else {
-		if (((icon_colors[settingvalue] - 0x0D) & 0xFE) == 0)	// if color == 0x0D or 0x0E
-			icon_colors[settingvalue] = 0x0F;
-	}	
-}
-
-void colordec() {
-	if (idx8_dec(icon_colors, settingvalue) & 0x30) {
-		if ((uint8_t)(icon_colors[settingvalue] & 0x0F) >= 0x0D)
-			idx8_store(icon_colors, settingvalue, (icon_colors[settingvalue] & 0x30) | 0x0C);
-	} else {
-		if (((icon_colors[settingvalue] - 0x0D) & 0xFE) == 0)	// if color == 0x0D or 0x0E
-			icon_colors[settingvalue] = 0x0C;
-	}
-}
